refactor(v2): move rust tests out of implementation files
