+++ programs/omnipair-v2/src/instructions/market/set_reduce_only.rs
@@ -58,10 +58,5 @@ impl<'info> SetMarketReduceOnly<'info> {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-
-    #[test]
-    fn set_market_reduce_only_uses_v1_emergency_authority() {
-        assert_ne!(REDUCE_ONLY_EMERGENCY_AUTHORITY, Pubkey::default());
-    }
+    include!("../../tests/instructions/market/set_reduce_only.rs");
 }
