+++ programs/omnipair-v2/src/tests/instructions/market/set_reduce_only.rs
@@ -1,0 +1,6 @@
+use super::*;
+
+    #[test]
+    fn set_market_reduce_only_uses_v1_emergency_authority() {
+        assert_ne!(REDUCE_ONLY_EMERGENCY_AUTHORITY, Pubkey::default());
+    }
